YTDB-510: Remove redundant type casts in 5 files (Qodana)
Remove 7 redundant casts where the expression already returns the
target type. Also remove unused imports that became unnecessary.

+++ core/src/main/java/com/jetbrains/youtrackdb/internal/core/metadata/schema/SchemaClassEmbedded.java
@@ -589,7 +589,7 @@ public class SchemaClassEmbedded extends SchemaClassImpl {
     if (name.toLowerCase(Locale.ENGLISH).equals(session.getCollectionNameById(collectionId))) {
       // DROP THE DEFAULT COLLECTION CALLED WITH THE SAME NAME ONLY IF EMPTY
       if (session.computeInTxInternal(
-          tx -> (long) session.countCollectionElements(collectionId)) == 0) {
+          tx -> session.countCollectionElements(collectionId)) == 0) {
         session.dropCollectionInternal(collectionId);
       }
     }
